Modified DOM Selectors and Cleaned Excess Elements
The previous query selector would insert 3x-4x the required requester
review overview elements. I narrowed the selector with a 2nd .forEach
loop so they're only added to the .requester-column portion of the row
now.

Also cleaned a bug that was causing 2 review elements to be shown in
smaller (mobile) viewports. Comparison can be found at:
http://prntscr.com/m133ki