Added Time Tracking & Return Reviews
Time Tracking: Added a small hourly/time tracker in place of MTurk's
default disaster of a countdown timer. Users can swap the display and
pause the hourly tracker same as TVJS. IMPORTANT NOTE: This is NOT
information that gets stored to MTS's HIT Tracker because, frankly,
indexedDB looks like a mess to my RDMS brain & I am VERY afraid of
breaking it. Will let Kada code review this & ask for his advice on
editing it or pray he pulls it in himself since I believe its just
adding the work_duration (int in ms, called from workDuration()) to
where Tracker stores its data but that whole area looks odd to me and I
don't want to experiment on live users where work history is concerned.

Return Reviews: These were added mostly for simplicity sake & because
they look nice in MTS as they're already pre-formatted for MTurks
default style guide. MTS is geared towards newer users who can use the
information more as well (it resembles TO's "this is bad" data mantra)
but functionally as a HUD its much more streamlined of an experience.
Also offer users a "safe" onboarding onto TV since they can submit "bad"
data (ie, claiming Endors are broken) and not hurt the overall system so
it should be a good bridge for both sides without adding a ton of
complexity right off the bat.

TV Notices: I've added quite a few heads up that TV's API is changing.
So difficult to balance not being annoying with the fact that workers
tend to ignore things until they break, feel free to make
edits/adjustments as needed. Finder has a TV modal explaining changes as
does the requester-reviews.js (so mostly places that call on TV's data)

TVJS Conflicts: MTS will now store a small DT variable in localstorage
that TVJS can read. TVJS will defer return reviews to MTS if the
variable exists in hopes that eventually TVJS can be sunset or
moved to a more "build & break" setup for staging/testing. After working
on MTS I want it to be an extension just for the icon in the nav bar of
Chrome haha. Either way this was the simplest solution I could find to
avoid duplicating requests & displays to users who have both installed.
Only known duplication is the hourly tracker as TVJS still needs its own
until full reviews are ported to v2 of the API (will wait for most users
to have had time for auth key generation before doing this to avoid bad
UX)

Manifest: I've never had to fiddle w/ manifest config before, it should
be correct but it is worth a once-over. I'm loading this as an unpacked
extension, unsure if a conflict would break on production but it "works
on my machine"